Workaround for the incorrect builders order
The correct order needs to be: 1) upstream repos builds, 2) PR repo
build and 3) the downstream build. However, since the upstream and
downstream builds are being manually configured, they ended up being
the first and second. So the PR repo build was executed _after_ the
downstream build, which makes no sense. Moreover the generated
WARs do not actually contain the changes from the PR, which is a big
issue. This is hopefully just a quick workaround before I find a better
way to fix it.