[BXMSPROD-161] Add cekit build capability to Jenkins image (#452)
This change adds dependencies and a script to the Jenkins image
which allows cekit osbs builds of rhba images. Since the items
necessary reference internal systems, the optional add-osbs.sh
script has been added to clone the necessary repository and install
the files. The rhba-osbs role and data directories have been stubbed
out for backward compatibility, and the README updated.

Signed-off-by: Trevor McKay <tmckay@redhat.com>